requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.6.15
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
constantly==23.10.4
cryptography==45.0.4
daphne==4.2.0
Django==4.2.20
django-cors-headers==4.7.0
django-mongoengine==0.5.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
dotenv==0.9.9
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
MarkupSafe==3.0.2
mongoengine==0.29.1
msgpack==1.1.0
packaging==24.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pymongo==4.11.3
PyMySQL==1.1.1
pyOpenSSL==25.1.0
python-dotenv==1.0.1
pytz==2025.1
redis==5.2.1
requests==2.32.3
service-identity==24.2.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==23.1.1
typing_extensions==4.14.0
urllib3==2.3.0
Werkzeug==3.1.3
whitenoise==6.9.0
zope.interface==7.2
